+++ create.go
@@ -2,16 +2,16 @@ package ap
 
 // Create instantiates an Assignment Problem of a given size.
 func Create(size int) *AP {
-	A := make([][]int, size) // 0
+	A := make([][]int64, size) // 0
 	for r := 0; r < size; r++ {
-		A[r] = make([]int, size)
+		A[r] = make([]int64, size)
 	}
 
 	f := make([]int, size)
 	fBar := make([]int, size)
 	p := make([]int, size)
 	c := make([]int, size)
-	pi := make([]int, size)
+	pi := make([]int64, size)
 
 	for i := 0; i < size; i++ {
 		f[i] = -1
@@ -23,8 +23,8 @@ func Create(size int) *AP {
 
 	return &AP{
 		A:    A,
-		U:    make([]int, size),
-		V:    make([]int, size),
+		U:    make([]int64, size),
+		V:    make([]int64, size),
 		f:    f,
 		fBar: fBar,
 		p:    p,
